backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
google-auth>=2.28
requests>=2.31
python-dotenv>=1.0
psycopg2-binary>=2.9
whitenoise>=6.6
gunicorn>=21.2
dj-database-url>=2.1